requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
mysql-connector-python==8.1.0
bcrypt==4.0.1
pandas==2.0.3
reportlab==4.0.4
python-dotenv==1.0.0
pillow==10.0.0